Stuff now compiles on Windows as well. We need to watch the undefined GetMessages in maidsafe-dht_config.h. Hopefully, the protocol buffer lads will change the name or solve the problem otherwise.
git-svn-id: https://maidsafe-dht.googlecode.com/svn/branches/signals@527 16f77258-3af3-11de-be05-5f7a86268029

+++ src/maidsafe/rpcprotocol/channelimpl.cc
@@ -184,7 +184,8 @@ void ChannelImpl::CallMethod(const google::protobuf::MethodDescriptor *method,
   }
 
   udt_connection_->Send(transport_message,
-                        pending_request.controller->timeout());
+                        static_cast<int>
+						    (pending_request.controller->timeout()));
 
 //  DLOG(INFO) << "Sent RPC request(" << rpc_message->method() << ") - "
 //             << socket_id << " to " << remote_ip_ << ":" << remote_port_
